% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-splits.R
\name{recycle-pipe}
\alias{recycle-pipe}
\alias{\%@\%}
\alias{\%>\%}
\title{Recycle dots pipe}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using the magrittr semantics.}
}
\description{
Pipe a recyclable object into a function with dots. Functionally identical to magrittr pipe.
}
