% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-splits.R
\name{t_test_split}
\alias{t_test_split}
\alias{chisq_split}
\alias{lm_split}
\title{Perform statistical tests w/'hacksaw-like' syntax}
\usage{
t_test_split(data, ..., args)

chisq_split(data, ..., args, tidier = c("tidy", "glance", "augment"))

lm_split(data, ..., args, tidier = c("tidy", "glance", "augment"))
}
\arguments{
\item{data}{a table of data}

\item{...}{possibly named formulas to pass on to tests}

\item{args}{a list of arguments to pass on to tests. Cf. ?t.test, ?chisq.test}

\item{tidier}{which tidier to use from the broom package}
}
\value{
a list
}
\description{
Perform statistical tests w/'hacksaw-like' syntax
}
\examples{
library(infer)
gss \%>\%
    t_test_split(
        hours ~ college,
        hours ~ sex
        )

gss \%>\%
    t_test_split(
        t1 = hours ~ college,
        t2 = hours ~ sex,
        args = list(alternative = "greater", conf.level = 0.9)
        )
}
